refactor: replace usage of ReleaseId from client with inline
ReleaseId from `@sanity/client` is deprecated as there is no guarantees of release ids starting with `r`
