+++ packages/sanity/src/core/comments/store/useCommentsStore.ts
@@ -1,12 +1,8 @@
-import {
-  type ListenEvent,
-  type ListenOptions,
-  type ReleaseId,
-  type SanityClient,
-} from '@sanity/client'
+import {type ListenEvent, type ListenOptions, type SanityClient} from '@sanity/client'
 import {useCallback, useEffect, useMemo, useReducer, useRef, useState} from 'react'
 import {catchError, of} from 'rxjs'
 
+import {type ReleaseId} from '../../perspective/types'
 import {getPublishedId} from '../../util'
 import {type CommentDocument, type Loadable} from '../types'
 import {commentsReducer, type CommentsReducerAction, type CommentsReducerState} from './reducer'
