+++ packages/sanity/src/core/form/inputs/ReferenceInput/types.ts
@@ -1,4 +1,3 @@
-import {type ReleaseId} from '@sanity/client'
 import {
   type I18nTextRecord,
   type Path,
@@ -9,6 +8,7 @@ import {
 import {type ComponentType, type ReactNode} from 'react'
 import {type Observable} from 'rxjs'
 
+import {type ReleaseId} from '../../../perspective/types'
 import {type DocumentAvailability} from '../../../preview'
 import {type PreviewState} from '../../../preview/utils/getPreviewStateObservable'
 import {type ObjectInputProps} from '../../types'
