+++ packages/sanity/src/core/perspective/navbar/GlobalPerspectiveMenu.tsx
@@ -1,4 +1,3 @@
-import {type ReleaseId} from '@sanity/client'
 import {ChevronDownIcon} from '@sanity/icons'
 // eslint-disable-next-line no-restricted-imports -- Button requires props, only supported by @sanity/ui
 import {Button, Menu} from '@sanity/ui'
@@ -7,6 +6,7 @@ import {styled} from 'styled-components'
 
 import {MenuButton} from '../../../ui-components'
 import {CreateReleaseDialog} from '../../releases/components/dialog/CreateReleaseDialog'
+import {type ReleaseId} from '../types'
 import {ReleasesList} from './ReleasesList'
 import {useScrollIndicatorVisibility} from './useScrollIndicatorVisibility'
