+++ packages/sanity/src/core/perspective/navbar/ReleaseTypeMenuSection.tsx
@@ -1,4 +1,3 @@
-import {type ReleaseId} from '@sanity/client'
 import {Flex, Label} from '@sanity/ui'
 import {useCallback} from 'react'
 
@@ -6,6 +5,7 @@ import {useTranslation} from '../../i18n/hooks/useTranslation'
 import {usePerspective} from '../../perspective/usePerspective'
 import {type ReleaseDocument, type ReleaseType} from '../../releases/store/types'
 import {getReleaseIdFromReleaseDocumentId} from '../../releases/util/getReleaseIdFromReleaseDocumentId'
+import {type ReleaseId} from '../types'
 import {
   getRangePosition,
   GlobalPerspectiveMenuItem,
