+++ packages/sanity/src/core/perspective/useSetPerspective.tsx
@@ -1,6 +1,7 @@
-import {type ReleaseId} from '@sanity/client'
 import {useCallback} from 'react'
 import {useRouter} from 'sanity/router'
+
+import {type ReleaseId} from './types'
 
 /**
  * @internal
