+++ packages/sanity/src/core/releases/hooks/__tests__/utils.test.ts
@@ -1,6 +1,6 @@
-import {type ReleaseId} from '@sanity/client'
 import {describe, expect, it} from 'vitest'
 
+import {type ReleaseId} from '../../../perspective/types'
 import {RELEASE_DOCUMENT_TYPE} from '../../store/constants'
 import {type ReleaseDocument} from '../../store/types'
 import {createReleaseId} from '../../util/createReleaseId'
