+++ packages/sanity/src/core/releases/hooks/useVersionOperations.tsx
@@ -1,8 +1,8 @@
-import {type ReleaseId} from '@sanity/client'
 import {useTelemetry} from '@sanity/telemetry/react'
 import {useToast} from '@sanity/ui'
 
 import {useTranslation} from '../../i18n'
+import {type ReleaseId} from '../../perspective/types'
 import {useSetPerspective} from '../../perspective/useSetPerspective'
 import {getDocumentVariantType} from '../../util/getDocumentVariantType'
 import {AddedVersion} from '../__telemetry__/releases.telemetry'
