+++ packages/sanity/src/structure/components/paneRouter/types.ts
@@ -1,6 +1,6 @@
-import {type ReleaseId} from '@sanity/client'
 import {type Path} from '@sanity/types'
 import {type ComponentType, type ReactNode} from 'react'
+import {type ReleaseId} from 'sanity'
 
 import {type RouterPanes, type RouterPaneSibling} from '../../types'
